world leaves always come first
nVisLeafs in the world model is a limit for linear rendering of the leaves array. If B's world leaves come after ALL of A' leaves then B's world leaves won't be rendered unless A has no submodel leaves.

This moves A's submodel leaves after B's leaves, so the order in the leaves array becomes:
- A's world leaves
- B's world leaves
- B's model leaves
- A's model leaves