added method to merge 3+ maps at once
maps are packed as equal-size blocks to form a bigger cube. The BSP tree will be deeper than necessary but limits will probably be hit long before that's a problem.

There's a big problem with lightmaps when moving maps. Coordinates don't get shifted and/or faces turn into a garbled mess. No idea how to fix this. The merged map can be relit with a compiler at least.

Also fixed ent bug.