progress on fixing lightmaps after moving a map
For whatever reason, the size of a face's lightmap depends not only on how big the face is, but also on where that face is in the world. I'm guessing this is caused by floating point precision errors. At least, that's the impression I get reading the hlrad code. Resizing lightmaps to match new face extents fixes the obvious glitches, but there are still seams between some faces if not resized correctly.

To resize, the original lightmap needs to be shifted somewhere inside the new canvas. It's not just a bilinear stretch of the original lightmap like you might expect. After literal days of guessing and debugging I found some code in hlrad that can predict how a lightmap would be shifted if it were compiled again. That added ~4k lines of voodoo-magic code here but hopefully most of that can be deleted after I figure out how the lightmap flags are calculated.

To validate the lightmap resizing, the map was recompiled with hlrad and the lightmaps were compared pixel-by-pixel. New pixels in a bigger canvas won't match, but those can be ignored. As long as the core of the lightmap is correct (everything with the "normal" lightmap flag), the seams disappear.

There's lots to be cleaned up now, but maybe some of this code can be used later. I'm not sure if minimizing the number of lightmap resizes is worth the effort, but that can sometimes be done by shifting texture coordinates or moving the map somewhere else. Lightmap resizing may be lossy (for big->small resizes), but that shouldn't be noticeable assuming the "shifted" pixels are never rendered.