light map fix progress
using lightmap luxel flags to figure out shifts is mostly accurate but sometimes the map is empty (all luxels "not lit") but has lighting anyway. So far this is only happening rarely and not on the world model. I give up trying to make this perfect. Models can probably be moved to prevent lightmaps being resized. Checking this in before ripping out any rad code that isn't needed.