huge speedup generating nav mesh
using an octree to reduce intersection tests for cutting, and only testing the center point of each poly for interior checks. Cut polys will either be wholly inside or outside of the level, so no need to check everywhere on the face. Also not using new split polygons to cut. A mesh for yabma generates in 3s, down from ~5 minutes.