nav mesh WIP
Improved splitting code and fixed some missing polys. Polygons are also merged and tiny polys stripped to reduce the node count. for yabma it makes a few hundred polys above what the recast lib makes (1700 -> 2000).

Outdoor areas with slopes end up looking very messy and have gaps. Using HULL 0 in those cases should help but it might be hard to know when that's possible, and it might not connect to HULL 3 mesh nicely.