leaf-based navigation mesh
got an idea to use navigation volumes for a plugin that tells players where to go. 3D volumes are simpler to link together than flat floor polygons with differing heights, and allow more pathing possibilities. The mesh is the set of empty leaves in the clipnode tree for a crouching player. Leaves are linked together by their overlapping face regions.