work around nvidia ion bug
it can't return mat4 from a function. The bone matrix has to be created in main().

While investigating this bug, much time was spent trying to use uniforms again, this time packing 16bit fixed point ints into an ivec[200]. It worked and packed just enough values for 1536 "floats" needed for 128 bone matrices while not exceeding the uniform component count. The problem was that I wanted this to run on GL 2.1 which doesn't have bitwise operators or unsigned ints. ints are also only guaranteed to be 16 bits and I need at least 28 or so for packing with enough precision. Another problem came from using floating point modulos to work around a gpu bug which prevents the normal modulo from working. The old gpu fights me every step of the way. I'm giving up. OpenGL 2.1 will have to use the old rendering method that's 10% as fast as doing vertex transforms on the gpu. Maybe meshes can be split up by groups of 40 bones or so and still be faster than CPU transforms.

The switch to GL_RGBA16F was made because GL_RGBA32F is not required by the gl 3.0 spec.