track worldleaves limit in Map Limits
A bug in the HL client limits the leaf counts to 8192 per model. There's a static buffer in Mod_LeafPVS that VIS decompresses to which is only 1024 bytes. At 8 leaves per byte, that's only 8192, not 32k.

This means new porting tools will be needed to merge leaves or move them to entity models.