world leaf reduction + leaf hiding
I couldn't get leaf merging to work. Merged leaves are sometimes invisible at certain angles, and it's not a PVS or bounding box problem. I think it might be occlusion using the node planes, somehow. If so, unlikely to fix without rebuilding the BSP tree.

"Convert to Model" converts selected leaves into a func_illusionary. Visibility and collision is preserved, but decals won't work. This is ok for unreachable areas and some nooks and crannies where players are unlikely to shoot. Another drawback is that if the func_illusionary is too big, it will be rendered everywhere in the map. This tool reduces world leaves by merging the selected leaves without preserving face visibility (the unfixable/difficult part of merging).

Leaf hiding will be used to isolate rooms for easy flood fill selection, for use with the func_illusionary tool.